# feat: implement comprehensive error handling system (issue #46)