## Summary

- Implement custom exception hierarchy with structured error context and timestamps
- Add error handler decorators for consistent file operations, validation, and performance monitoring
- Include security features with path sanitization and credential redaction for safe logging
- Provide graceful degradation patterns for non-critical operations using safe_operation decorator
- Integrate error handling system into MinervaService methods with proper validation and performance logging

## Technical Details

### Custom Exception Hierarchy
- **MinervaError**: Base exception with structured context and timestamps
- **ValidationError**: Input validation failures
- **NoteNotFoundError/NoteExistsError**: File operation specific errors
- **VaultError**: Permission and I/O issues
- **TagError/SearchError/FrontmatterError/ConfigurationError**: Domain-specific errors

### Error Handler Utilities
- **@handle_file_operations()**: Converts standard filesystem exceptions to Minerva exceptions
- **@validate_inputs()**: Input validation with consistent error handling
- **@log_performance()**: Automatic logging of slow operations with configurable thresholds
- **@safe_operation()**: Graceful error handling for non-critical operations

### Security Features
- Path sanitization in logs (vault paths shown as `<vault>/...`, external paths as `<external>/...`)
- Automatic credential redaction for sensitive fields (password, token, secret, key)
- Secure error context creation with sanitized information

### Performance Monitoring
- High-precision timing using `time.perf_counter()` for accurate measurement
- Configurable thresholds: 500ms for file creation/editing, 200ms for reading, 300ms for tags, 1000ms for search
- Automatic logging of operations exceeding thresholds

### Service Integration
- Applied decorators to key MinervaService methods (create_note, edit_note, read_note, add_tag, etc.)
- Performance thresholds optimized per operation type
- get_tags method uses safe_operation for graceful degradation

### CodeRabbit Review Improvements
- Added missing `@handle_file_operations()` decorator to edit_note method for consistency
- Removed unused `_validate_tag` function to eliminate name clash with instance method
- Fixed off-by-one error in test validator for accurate argument position
- Replaced `time.time()` with `time.perf_counter()` for more accurate performance measurement
- Updated test expectations from Exception to NoteNotFoundError for better error specificity

### Build System Fixes
- Added hatchling build backend to pyproject.toml
- Set tool.uv.package = true for proper package recognition
- Updated Makefile with PYTHONPATH=src for all commands to fix module import issues

### Type Safety
- Resolved all mypy type checking errors with proper type annotations
- Added cast() to decorator return types for TypeVar compatibility
- Enhanced type safety for flexible input handling in error context creation

### Documentation
- Created comprehensive Japanese documentation in docs/error_handling.md
- Added language usage rules to CLAUDE.md from .github/instructions/language_rules.md
- Included migration guide, testing examples, and best practices
- Updated technical details on timing accuracy and decorator consistency

## Test Coverage

- **tests/test_exceptions.py**: Custom exception behavior and context
- **tests/test_error_handler.py**: Error handler utilities and decorators
- **tests/test_service_error_handling.py**: Service layer integration
- Updated existing integration tests for new exception types
- All 264 tests pass successfully with comprehensive coverage

## Breaking Changes

None - the error handling system is designed for backward compatibility while improving error classification and logging.

## Migration Guide

Existing error handling patterns continue to work, but new patterns are recommended:

```python
# Old pattern
try:
    result = some_operation()
except Exception as e:
    logger.error("操作が失敗しました: %s", e)
    raise

# New pattern (automatic with decorators)
@handle_file_operations()
@log_performance()
def some_operation():
    # Automatic error handling and logging
    pass
```

Exception mapping:
- `ValueError` → `ValidationError` (for input validation)
- `FileNotFoundError` → `NoteNotFoundError`
- `FileExistsError` → `NoteExistsError`
- `PermissionError` → `VaultError`
- `IOError`/`OSError` → `VaultError`

🤖 Generated with [Claude Code](https://claude.ai/code)